requirements.txt
altair==5.5.0
anndata==0.11.4
annotated-types==0.7.0
array-api-compat==1.12.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h5py==3.14.0
httplib2==0.22.0
idna==3.10
igraph==0.11.9
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
leidenalg==0.10.2
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.42.1
natsort==8.4.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
plotly==6.1.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
rsa==4.9.1
scanpy==1.11.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
session-info2==0.1.2
six==1.17.0
smmap==5.0.2
statsmodels==0.14.4
streamlit==1.45.1
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.7
uritemplate==4.2.0
urllib3==2.4.0
